Supported build infrastruture using gclient and gn.
These tools are used and contributed by Chromium opensource project.

depottools - http://www.chromium.org/developers/how-tos/depottools

gn - https://code.google.com/p/chromium/wiki/gn

+++ sorting/insertion_sort/1.cc
@@ -7,7 +7,7 @@
 namespace {
 template <class T>
 void insertion_sort(std::vector<T>& input) {
-  for(int i = 1; i < input.size(); i++) {
+  for(size_t i = 1; i < input.size(); i++) {
     auto key = input[i];
     int p = i - 1;
     while(p >= 0 && key < input[p]) {
